+++ src/ezdxf/entities/xdata.py
@@ -41,7 +41,7 @@ class XData:
 
     def add(self, appid: str, tags: Iterable) -> None:
         data = Tags(dxftag(code, value) for code, value in tags)
-        if data[0] != (XDATA_MARKER, appid):
+        if len(data) == 0 or data[0] != (XDATA_MARKER, appid):
             data.insert(0, dxftag(XDATA_MARKER, appid))
         self._add(data)
 
@@ -49,7 +49,7 @@ class XData:
         if appid in self.data:
             return self.data[appid]
         else:
-            raise DXFKeyError(appid)
+            raise DXFValueError(appid)
 
     def discard(self, appid):
         if appid in self.data:
@@ -126,7 +126,7 @@ class XData:
         """
         try:
             xdata = self.get(appid)
-        except DXFKeyError:
+        except DXFValueError:
             pass
         else:
             try: