+++ NEWS.md
@@ -7,6 +7,7 @@ Version 0.12.2 - dev
 
 - BUGFIX: `XData.get()` now raises `DXFValueError` for not existing appids, like all other methods of the `XData()` class
 - BUGFIX: `Layer.description` returns an empty string for unknown XDATA structure in `AcAecLayerStandard`
+- BUGFIX: Initialize/Load `Hatch` edge coordinates as `Vec2()` objects
 
 Version 0.12.1 - 2020-04-25
 ---------------------------